Exp7/requirements.txt
dvc
scikit-learn
seaborn
pandas
numpy
pyyaml
uvicorn
